More champion-related tests
I'm defaulting to comparing the JSON generated by the RiotWatcher
methods to the JSON from the correct URLs. This avoids the problem of
just comparing the URLs, both of which may be wrong, and the problem of
getting the methods' URLs at all, since there's no method for that.
Maybe there should be.